Extracts models from a mediation object as a data frame

When conducting a joint significant test, different models are fitted to the data. This function helps you access the models used in an object of class mediation_model.

extract_tidy_models(mediation_model)

Arguments

mediation_model
An object of class mediation_model.

Value

A data frame.

See also

Examples

data(ho_et_al)
ho_et_al$condition_c <- build_contrast(ho_et_al$condition,
                                       "Low discrimination",
                                       "High discrimination")
my_model <-
  mdt_simple(data = ho_et_al,
             IV = condition_c,
             DV = hypodescent,
             M = linkedfate)

extract_tidy_models(my_model)

#> # A tibble: 7 x 6
#>   model      term        estimate std.error statistic  p.value
#>   <chr>      <chr>          <dbl>     <dbl>     <dbl>    <dbl>
#> 1 X -> Y     (Intercept)   4.57      0.0403   113.    0       
#> 2 X -> Y     condition_c   0.171     0.0805     2.13  3.35e- 2
#> 3 X -> M     (Intercept)   5.16      0.0424   122.    0       
#> 4 X -> M     condition_c   0.772     0.0848     9.10  6.53e-19
#> 5 X + M -> Y (Intercept)   3.60      0.172     20.9   3.09e-78
#> 6 X + M -> Y condition_c   0.0273    0.0829     0.329 7.42e- 1
#> 7 X + M -> Y linkedfate    0.187     0.0325     5.75  1.27e- 8